feat: Add `useActiveAccount` hook
