+++ apps/tangle-dapp/hooks/useSubstrateInjectedExtension.ts
@@ -1,21 +1,20 @@
 import { InjectedExtension } from '@polkadot/extension-inject/types';
+import { useActiveAccount } from '@webb-tools/api-provider-environment/hooks/useActiveAccount';
+import usePromise from '@webb-tools/tangle-shared-ui/hooks/usePromise';
 import { findInjectorForAddress } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import { useCallback } from 'react';
 
-import useActiveAccountAddress from './useActiveAccountAddress';
-import usePromise from './usePromise';
-
 export default function useSubstrateInjectedExtension() {
-  const activeAccountAddress = useActiveAccountAddress();
+  const [activeAccount] = useActiveAccount();
 
   const { result: injector } = usePromise<InjectedExtension | null>(
     useCallback(() => {
-      if (activeAccountAddress === null) {
+      if (activeAccount === null) {
         return Promise.resolve(null);
       }
 
-      return findInjectorForAddress(activeAccountAddress);
-    }, [activeAccountAddress]),
+      return findInjectorForAddress(activeAccount.address);
+    }, [activeAccount]),
     null,
   );
