+++ libs/api-provider-environment/src/WebbProvider/subjects.ts
@@ -1,4 +1,3 @@
-import type { Account } from '@webb-tools/abstract-api-provider';
 import type { Chain, WalletConfig } from '@webb-tools/dapp-config';
 import { Maybe, Nullable } from '@webb-tools/dapp-types/utils/types';
 import { useObservableState } from 'observable-hooks';
@@ -17,19 +16,7 @@ export function useActiveWallet() {
   return [activeWallet, setActiveWallet] as const;
 }
 
-/** Active account subject */
-const activeAccountSubject = new BehaviorSubject<Account | null>(null);
-
-/** Active account setter */
-const setActiveAccount = (account: Account | null) =>
-  activeAccountSubject.next(account);
-
 /** Hook for sharing the active account subject & setter */
-export function useActiveAccount() {
-  const activeAccount = useObservableState(activeAccountSubject);
-  return [activeAccount, setActiveAccount] as const;
-}
-
 /**
  * Active chain subject
  * - `undefined` means no chain is active
