+++ libs/api-provider-environment/src/hooks/useActiveAccount.ts
@@ -1,0 +1,15 @@
+import type { Account } from '@webb-tools/abstract-api-provider';
+import { useObservableState } from 'observable-hooks';
+import { BehaviorSubject } from 'rxjs';
+
+/** Active account subject */
+const activeAccountSubject = new BehaviorSubject<Account | null>(null);
+
+/** Active account setter */
+const setActiveAccount = (account: Account | null) =>
+  activeAccountSubject.next(account);
+
+export function useActiveAccount() {
+  const activeAccount = useObservableState(activeAccountSubject);
+  return [activeAccount, setActiveAccount] as const;
+}
